feat: update wiremock - [x] use wiremock junit5 - [x] refactor base wiremock setup - [x] update integrationTest setup - [x] update wire-mock package
