+++ shared/utils/build.gradle
@@ -1,7 +1,7 @@
 apply from: "$rootDir/dependency/jackson.gradle"
 
 dependencies {
-    testImplementation libs.bundles.wiremock
+    testImplementation libs.wiremock
     testImplementation libs.rest.assured
     testImplementation project(":shared:mock")
 }
