+++ springboot/wire-mock/src/main/java/com/github/durex/Application.java
@@ -1,11 +1,12 @@
-package com.github.durex.todo;
+package com.github.durex;
 
+import lombok.Generated;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class Application {
-
+  @Generated
   public static void main(String[] args) {
     SpringApplication.run(Application.class, args);
   }
