+++ springboot/wire-mock/src/main/java/com/github/durex/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.github.durex.todo;
+package com.github.durex.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
