+++ springboot/wire-mock/src/main/java/com/github/durex/config/WebClientConfig.java
@@ -1,4 +1,4 @@
-package com.github.durex.todo;
+package com.github.durex.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
