+++ springboot/wire-mock/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+todo_base_url= https://jsonplaceholder.typicode.com
