--- springboot/wire-mock/src/main/resources/application.yml
@@ -1,1 +1,0 @@
-todo_base_url: https://jsonplaceholder.typicode.com
