--- springboot/wire-mock/src/test/java/com/github/durex/todo/WireMockInitializer.java
@@ -1,33 +1,0 @@
-package com.github.durex.todo;
-
-import com.github.tomakehurst.wiremock.WireMockServer;
-import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
-import java.util.Map;
-import org.springframework.boot.test.util.TestPropertyValues;
-import org.springframework.context.ApplicationContextInitializer;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.event.ContextClosedEvent;
-
-public class WireMockInitializer
-    implements ApplicationContextInitializer<ConfigurableApplicationContext> {
-
-  @Override
-  public void initialize(ConfigurableApplicationContext applicationContext) {
-
-    WireMockServer wireMockServer = new WireMockServer(new WireMockConfiguration().dynamicPort());
-
-    wireMockServer.start();
-
-    applicationContext.addApplicationListener(
-        applicationEvent -> {
-          if (applicationEvent instanceof ContextClosedEvent) {
-            wireMockServer.stop();
-          }
-        });
-
-    applicationContext.getBeanFactory().registerSingleton("wireMockServer", wireMockServer);
-
-    TestPropertyValues.of(Map.of("todo_base_url", wireMockServer.baseUrl()))
-        .applyTo(applicationContext);
-  }
-}
