+++ springboot/wire-mock/src/test/resources/application-test.properties
@@ -1,0 +1,1 @@
+todo_base_url= http://127.0.0.1:8080
